@@ -88,7 +88,7 @@ export class SelectAutocompleteComponent implements OnChanges, DoCheck {
   @Output()
   selectionChange: EventEmitter<any> = new EventEmitter();
 
-  @ViewChild('selectElem') selectElem;
+  @ViewChild('selectElem', { static: false }) selectElem;
 
   filteredOptions: Array<any> = [];
   selectedValue: Array<any> = [];
