@@ -7,7 +7,8 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatIconModule } from '@angular/material/icon';
-import { SelectAutocompleteModule } from 'select-autocomplete';
+// import { SelectAutocompleteModule } from 'select-autocomplete';
+import { SelectAutocompleteModule } from './../../projects/select-autocomplete/src/lib/select-autocomplete.module';
 
 import { AppComponent } from './app.component';
 
